requirements.txt
annotated-types==0.7.0
anyio==4.11.0
bcrypt==4.0.1
click==8.3.0
colorama==0.4.6
dnspython==2.8.0
exceptiongroup==1.3.0
fastapi==0.119.0
h11==0.16.0
idna==3.11
passlib==1.7.4
pydantic==2.12.2
pydantic_core==2.41.4
PyJWT==2.10.1
pymongo==4.15.3
python-dotenv==1.1.1
python-multipart==0.0.20
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
